requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
groq==0.8.0
pydantic==2.6.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0